Add grace period for turning off fan
The original script just turned the fan on or off
depending on if the current temp was above some
threshold. Around that threshold, the fan could turn
on and off repeatedly as the temp rose slightly above
and below.

This new code adds a 10 minute grace period such that
after the fan turns on, the code will wait 10 minutes
before turning off the fan, regardless of if the temp
is below the established threshold.